+++ src/components/atoms/Card/Card.tsx
@@ -3,11 +3,12 @@ import './Card.scss';
 
 interface CardProps {
   children: React.ReactNode;
+  className?: string;
 }
 
-const Card: React.FC<CardProps> = ({ children }) => {
+const Card: React.FC<CardProps> = ({ children, className }) => {
   return (
-    <div className="card">
+    <div className={`card ${className}`}>
       {children}
     </div>
   );
