+++ src/pages/Login/Login.scss
@@ -1,0 +1,20 @@
+.login-page {
+  .section {
+    height: 100%;
+
+    .section-container {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+    }
+  }
+
+  &__card {
+    max-width: 380px;
+  }
+
+  &__logo {
+    margin: 0 auto;
+    width: fit-content;
+  }
+}
